air quality api controller
